// frontend/mobilephone-frontend/src/types/phone.ts
export interface Phone {
  id: number;
  brand: string;
  model: string;
  storage: string;
  ram: string;
  screenSize: string;
  camera: string;
  battery: string;
  imageUrl: string;
  weight?: number;
  dimensions?: string;
  processor?: string;
  os?: string;
  releaseYear?: number;
  networkType?: string;
  chargingPower?: string;
  waterResistance?: string;
  material?: string;
  colors?: string;
  colorImages?: string;
  imageFront?: string;
  imageBack?: string;
  imageSide?: string;
}